requirements.txt
Django==1.11.6
amqp==2.2.2
billiard==3.5.0.3
celery[redis]==4.1.0
djangorestframework==3.7.1
kombu==4.1.0
pytz==2017.2
redis==2.10.6
vine==1.1.4
drf-writable-nested==0.3.1
